fix: rename files to use kebab-case
